jsonld: Add proposed "lite" schema
Adds a proposed method for implementing a "lite" JSON-LD SPDX format.
Instead of re-specifying the spec, the lite profile is instead a small
hand-curated JSON schema that simply forces the root object of the
document to be a SpdxDocument. This works fine because the SpdxDocument
is already supposed to encompass the document elements in its elements
array, so inlining them makes sense anyway.

This means that the "lite" document is also completely valid JSON-LD,
which means it can be validated by the SHACL model, or even the full
JSON-LD schema.

It also means it automatically keeps up with the main schema document
when it changes